Move the HeaderMenuDropDownStyles to their component
