+++ src/connectors/components/HeaderMenuExchangeConnector.js
@@ -7,8 +7,7 @@ import { exchangeMax } from '../../actions/CryptoExchangeActions.js'
 import { showHelpModal } from '../../components/modals/HelpModal.js'
 import * as Constants from '../../constants/indexConstants'
 import s from '../../locales/strings.js'
-import { type DispatchProps, type StateProps, MenuDropDown } from '../../modules/UI/components/MenuDropDown/MenuDropDown.ui.js'
-import { MenuDropDownStyle } from '../../styles/components/HeaderMenuDropDownStyles.js'
+import { type DispatchProps, type StateProps, MenuDropDown, MenuDropDownStyle } from '../../modules/UI/components/MenuDropDown/MenuDropDown.ui.js'
 import THEME from '../../theme/variables/airbitz'
 import type { Dispatch, State } from '../../types/reduxTypes.js'
