+++ src/connectors/components/HeaderMenuRequestConnector.js
@@ -5,8 +5,7 @@ import { connect } from 'react-redux'
 import { showHelpModal } from '../../components/modals/HelpModal.js'
 import * as Constants from '../../constants/indexConstants'
 import s from '../../locales/strings.js'
-import { MenuDropDown } from '../../modules/UI/components/MenuDropDown/MenuDropDown.ui.js'
-import { MenuDropDownStyle } from '../../styles/components/HeaderMenuDropDownStyles.js'
+import { MenuDropDown, MenuDropDownStyle } from '../../modules/UI/components/MenuDropDown/MenuDropDown.ui.js'
 import THEME from '../../theme/variables/airbitz'
 import type { Dispatch, State } from '../../types/reduxTypes.js'
